api/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
asyncpg
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=4.0.0,<5.0.0
python-multipart
python-dotenv
httpx
pytest
pytest-asyncio
